bigInt returned as int now url format change

+++ prisma/sql/getProviderBiggestClientDistribution.sql
@@ -5,7 +5,7 @@ with providers_with_ratio as (select provider,
                               group by provider, week)
 select 100 * ceil(biggest_to_total_ratio::float8 * 20) / 20 - 5 as "value_from_exclusive",
        100 * ceil(biggest_to_total_ratio::float8 * 20) / 20     as "value_to_inclusive",
-       count(*)                                                 as "count",
+       count(*)::int as "count",
        week
 from providers_with_ratio
 group by "value_from_exclusive", "value_to_inclusive", week
